refactor: separate OS-specific config path logic by platform
