+++ internal/backdrop.go
@@ -69,16 +69,24 @@ func BackdropAction(out io.Writer, config *Config, args []string) error {
 	return nil
 }
 
-func handleSelectionConfirmation(previousWallpaper string, out io.Writer, cleanup func()) (bool, error) {
+func handleSelectionConfirmation(previousWallpaper, prompt, successMessage string, out io.Writer, cleanup func()) (bool, error) {
+	if prompt == "" {
+		prompt = "Want to save this change? [y/N]: "
+	}
+
+	if successMessage == "" {
+		successMessage = "Successfully changed background image!"
+	}
+
 	for {
-		userInput, err := userConfirmation(inputConfirmation)
+		userInput, err := userConfirmationWithPrompt(inputConfirmation, prompt)
 		if err != nil {
 			return false, err
 		}
 
 		switch userInput {
 		case "y":
-			fmt.Fprintln(out, "Successfully changed background image!")
+			fmt.Fprintln(out, successMessage)
 			return true, nil
 		case "n", "":
 			if err := setWallpaper(previousWallpaper); err != nil {
@@ -93,9 +101,9 @@ func handleSelectionConfirmation(previousWallpaper string, out io.Writer, cleanu
 	}
 }
 
-func userConfirmation(r io.Reader) (string, error) {
+func userConfirmationWithPrompt(r io.Reader, prompt string) (string, error) {
 	reader := bufio.NewReader(r)
-	fmt.Print("Want to save this change? [y/N]: ")
+	fmt.Print(prompt)
 
 	input, err := reader.ReadString('\n')
 	if err != nil {
