+++ internal/fuzzy.go
@@ -46,7 +46,7 @@ func handleFuzzySearch(out io.Writer, wallpapersPath string, wallpapers []string
 			return err
 		}
 
-		hasConfirmed, err = handleSelectionConfirmation(previousWallpaper, out, func() {})
+		hasConfirmed, err = handleSelectionConfirmation(previousWallpaper, "", "", out, func() {})
 		if err != nil {
 			return err
 		}
