+++ test/slideshow-original.go
@@ -1,0 +1,386 @@
+package internal
+
+import (
+	"bufio"
+	"fmt"
+	"io"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"strconv"
+	"strings"
+)
+
+var (
+	inputDuration io.Reader = os.Stdin
+)
+
+func handleSlideshow(out io.Writer, wallpapersPath string, wallpapers []string, imageSelection FuzzySelection) error {
+	for hasConfirmed := false; !hasConfirmed; {
+		previousWallpaper, err := getPreviousWallpaper()
+		if err != nil {
+			return err
+		}
+
+		selectedWallpaper, err := imageSelection(wallpapers)
+		if err != nil {
+			return err
+		}
+
+		duration, err := userDuration(inputDuration)
+		if err != nil {
+			return err
+		}
+
+		selectedWallpaper, err = configureSlideShow(selectedWallpaper, wallpapersPath, duration)
+		if err != nil {
+			return err
+		}
+
+		if runtime.GOOS != "windows" {
+			err = setWallpaper(selectedWallpaper)
+			if err != nil {
+				return err
+			}
+		}
+
+		hasConfirmed, err = handleSelectionConfirmation(previousWallpaper, out, func() {})
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func userDuration(r io.Reader) (int, error) {
+	reader := bufio.NewReader(r)
+	fmt.Print("What should be the duration per slide? (In minutes): ")
+
+	input, err := reader.ReadString('\n')
+	if err != nil {
+		return 0, fmt.Errorf("failed to read duration input: %w", err)
+	}
+
+	input = strings.TrimSpace(input)
+	duration, err := strconv.Atoi(input)
+	if err != nil || duration <= 0 {
+		return 0, fmt.Errorf("invalid duration: please enter a positive integer")
+	}
+	duration = duration * 60_000
+	return duration, nil
+}
+
+func configureSlideShow(imageText, wallpapersPath string, duration int) (string, error) {
+	images := strings.Split(imageText, ";")
+
+	switch runtime.GOOS {
+	case "linux":
+		return configureSlideShowLinux(images, wallpapersPath, duration)
+	case "windows":
+		return configureSlideShowWindows(images, wallpapersPath, duration)
+
+	default:
+		return "", ErrNoCompatibleOS
+	}
+}
+
+func configureSlideShowWindows(images []string, wallpapersPath string, duration int) (string, error) {
+	slideShowDir := filepath.Join(os.Getenv("APPDATA"), "BackdropSlideShow")
+
+	if err := os.MkdirAll(slideShowDir, 0777); err != nil {
+		return "", fmt.Errorf("failed to create slideshow directory: %w", err)
+	}
+
+	for _, img := range images {
+		src := filepath.Join(wallpapersPath, img)
+		dst := filepath.Join(slideShowDir, filepath.Base(img))
+
+		data, err := os.ReadFile(src)
+
+		if err != nil {
+			return "", fmt.Errorf("failed to read image %s: %w", src, err)
+		}
+
+		if err := os.WriteFile(dst, data, 0666); err != nil {
+			return "", fmt.Errorf("failed to write imageto slideshow folder: %w", err)
+		}
+	}
+
+	if err := setWindowsSlideshow(slideShowDir, duration); err != nil {
+		return "", err
+	}
+	firstImagePath := filepath.Join(slideShowDir, filepath.Base(images[0]))
+	themeFile, err := createWindowsThemeFile(firstImagePath, slideShowDir, duration)
+	if err != nil {
+		return "", fmt.Errorf("failed to create theme file: %w", err)
+	}
+
+	err = applyWindowsTheme(themeFile)
+	if err != nil {
+		return "", err
+	}
+
+	return slideShowDir, nil
+}
+
+func setWindowsSlideshow(folder string, duration int) error {
+
+	cmd := exec.Command("powershell", "-Command", fmt.Sprintf(`
+$RegPath = "HKCU:\Control Panel\Personalization\Desktop Slideshow"
+Set-ItemProperty -Path $RegPath -Name Interval -Value %d
+Set-ItemProperty -Path $RegPath -Name Shuffle -Value 1
+Set-ItemProperty -Path $RegPath -Name SlideshowEnabled -Value 1
+
+$ThemePath = "HKCU:\Software\Microsoft\Windows\CurrentVersion\Themes"
+Set-ItemProperty -Path $ThemePath -Name SlideshowDirectory -Value "%s"
+
+$WallpaperPath = "HKCU:\Control Panel\Desktop"
+Set-ItemProperty -Path $WallpaperPath -Name Wallpaper -Value ""
+Set-ItemProperty -Path $WallpaperPath -Name WallpaperStyle -Value 10
+Set-ItemProperty -Path $WallpaperPath -Name TileWallpaper -Value 0
+
+# Critical: Remove corrupted or cached wallpaper to avoid black background
+$transcoded = "$env:APPDATA\Microsoft\Windows\Themes\TranscodedWallpaper"
+if (Test-Path $transcoded) { Remove-Item $transcoded -Force -ErrorAction SilentlyContinue }
+
+RUNDLL32.EXE user32.dll, UpdatePerUserSystemParameters
+`, duration, folder))
+
+	if output, err := cmd.CombinedOutput(); err != nil {
+		return fmt.Errorf("failed to set registry slideshow values: %v\n%s", err, output)
+	}
+
+	return nil
+}
+
+func configureSlideShowLinux(images []string, wallpapersPath string, duration int) (string, error) {
+
+	slideShowFile, slideShowConfigFile, err := createSlideShowDirectory()
+	if err != nil {
+		return "", err
+	}
+
+	if err := createSlideShowFile(slideShowFile, slideShowConfigFile); err != nil {
+		return "", err
+	}
+
+	slideShowConfigFile, err = createSlideShowConfigFile(images, slideShowConfigFile, wallpapersPath, duration)
+	if err != nil {
+		return "", err
+	}
+
+	return slideShowConfigFile, nil
+}
+
+func createSlideShowDirectory() (string, string, error) {
+	switch runtime.GOOS {
+	case "linux":
+		homePath, err := os.UserHomeDir()
+		if err != nil {
+			return "", "", fmt.Errorf("unable to find user home directory: %w", err)
+		}
+		return configureLinuxSlideShowPaths(homePath)
+	case "windows":
+		return configureWindowsSlideShowPaths()
+
+	default:
+		return "", "", ErrNoCompatibleOS
+	}
+}
+
+func configureWindowsSlideShowPaths() (string, string, error) {
+	return "", "", fmt.Errorf("configureWindowsSlideShowPaths is not implemented")
+}
+
+func configureLinuxSlideShowPaths(homePath string) (string, string, error) {
+	paths := map[string]string{
+		"slideShowPath":       filepath.Join(homePath, ".local", "share", "gnome-background-properties"),
+		"slideShowConfigPath": filepath.Join(homePath, ".local", "share", "backgrounds", "backdrop_settings"),
+	}
+
+	for _, path := range paths {
+		if err := os.MkdirAll(path, 0777); err != nil {
+			return "", "", fmt.Errorf("failed to create directory %s: %w", path, err)
+		}
+	}
+
+	slideShowFile := filepath.Join(paths["slideShowPath"], "backdrop_slideshow.xml")
+	slideShowConfigFile := filepath.Join(paths["slideShowConfigPath"], "backdrop_settings.xml")
+
+	return slideShowFile, slideShowConfigFile, nil
+}
+
+func createSlideShowFile(outFile, configFile string) error {
+	var content strings.Builder
+
+	content.WriteString(`<?xml version="1.0" encoding="UTF-8"?>`)
+	content.WriteString("\n")
+	content.WriteString(`<!DOCTYPE wallpapers SYSTEM "gnome-wp-list.dtd">`)
+	content.WriteString("\n")
+	content.WriteString("<wallpapers>\n")
+	content.WriteString("  <wallpaper>\n")
+	content.WriteString("    <name>Backdrop Slideshow</name>\n")
+	content.WriteString(fmt.Sprintf("     <filename>%v</filename>\n", configFile))
+	content.WriteString("    <options>zoom</options>\n")
+	content.WriteString("    <pcolor>#2c001e</pcolor>\n")
+	content.WriteString("    <scolor>#2c001e</scolor>\n")
+	content.WriteString("    <shade_type>solid</shade_type>\n")
+	content.WriteString("  </wallpaper>\n")
+	content.WriteString("</wallpapers>\n")
+
+	file, err := os.Create(outFile)
+	if err != nil {
+		return fmt.Errorf("failed to create slideshow file at %s: %w", outFile, err)
+	}
+	defer file.Close()
+
+	if _, err := file.WriteString(content.String()); err != nil {
+		return fmt.Errorf("failed to write content to slideshow file %s: %w", outFile, err)
+	}
+
+	if err := os.Chmod(outFile, 0777); err != nil {
+		return fmt.Errorf("failed to set permissions for slideshow file %s: %w", outFile, err)
+	}
+
+	return nil
+}
+
+func createSlideShowConfigFile(images []string, configFile, wallpapersPath string, duration int) (string, error) {
+	var content strings.Builder
+
+	content.WriteString("<background>\n")
+	content.WriteString("  <starttime>\n")
+	content.WriteString("    <year>2012</year>\n")
+	content.WriteString("    <month>01</month>\n")
+	content.WriteString("    <day>01</day>\n")
+	content.WriteString("    <hour>00</hour>\n")
+	content.WriteString("    <minute>00</minute>\n")
+	content.WriteString("    <second>00</second>\n")
+	content.WriteString("  </starttime>\n")
+
+	for i := 0; i < len(images)-1; i++ {
+		currentImage := filepath.Join(wallpapersPath, images[i])
+		nextImage := filepath.Join(wallpapersPath, images[i+1])
+		content.WriteString("  <static>\n")
+		content.WriteString(fmt.Sprintf("   <duration>%d.0</duration>\n", duration))
+		content.WriteString(fmt.Sprintf("   <file>%v</file>\n", currentImage))
+		content.WriteString("  </static>\n")
+		content.WriteString("   <transition>\n")
+		content.WriteString("   <duration>0.5</duration>\n")
+		content.WriteString(fmt.Sprintf("   <from>%v</from>\n", currentImage))
+		content.WriteString(fmt.Sprintf("   <to>%v</to>\n", nextImage))
+		content.WriteString("   </transition>\n")
+	}
+
+	startImage := filepath.Join(wallpapersPath, images[0])
+	endImage := filepath.Join(wallpapersPath, images[len(images)-1])
+	content.WriteString("  <static>\n")
+	content.WriteString(fmt.Sprintf("   <duration>%d.0</duration>\n", duration))
+	content.WriteString(fmt.Sprintf("   <file>%v</file>\n", endImage))
+	content.WriteString("  </static>\n")
+	content.WriteString("  <transition>\n")
+	content.WriteString("   <duration>0.5</duration>\n")
+	content.WriteString(fmt.Sprintf("   <from>%v</from>\n", endImage))
+	content.WriteString(fmt.Sprintf("   <to>%v</to>\n", startImage))
+	content.WriteString("  </transition>\n")
+	content.WriteString("</background>\n")
+
+	file, err := os.Create(configFile)
+	if err != nil {
+		return "", fmt.Errorf("failed to create slideshow config file at %s: %w", configFile, err)
+	}
+	defer file.Close()
+
+	if _, err := file.WriteString(content.String()); err != nil {
+		return "", fmt.Errorf("failed to write content to config file %s: %w", configFile, err)
+	}
+
+	if err := os.Chmod(configFile, 0777); err != nil {
+		return "", fmt.Errorf("failed to set permissions for config file %s: %w", configFile, err)
+	}
+
+	return file.Name(), nil
+}
+
+func createWindowsThemeFile(firstImagePath, slideshowDir string, duration int) (string, error) {
+	themesDir := filepath.Join(os.Getenv("LOCALAPPDATA"), "Microsoft", "Windows", "Themes")
+	if err := os.MkdirAll(themesDir, 0777); err != nil {
+		return "", fmt.Errorf("failed to create themes directory: %w", err)
+	}
+
+	themeFilePath := filepath.Join(themesDir, "backdrop.theme")
+
+	content := fmt.Sprintf(`[Theme]
+DisplayName=Backdrop Slideshow
+
+[Control Panel\Desktop]
+wallpaper=%s
+TileWallpaper=0
+WallpaperStyle=10
+PicturePosition=10
+SlideshowEnabled=1
+MultimonBackgrounds=1
+
+[Slideshow]
+ImagesRootPath=%s
+Interval=%d
+Shuffle=1
+
+[VisualStyles]
+Path=%%SystemRoot%%\resources\Themes\Aero\Aero.msstyles
+ColorStyle=NormalColor
+Size=NormalSize
+AutoColorization=0
+VisualStyleVersion=10
+
+[MasterThemeSelector]
+MTSM=RJSPBS
+
+[Sounds]
+SchemeName=@mmres.dll,-800
+`, firstImagePath, slideshowDir, duration)
+
+	if err := os.WriteFile(themeFilePath, []byte(content), 0666); err != nil {
+		return "", fmt.Errorf("failed to write theme file: %w", err)
+	}
+
+	return themeFilePath, nil
+}
+
+func applyWindowsTheme(themePath string) error {
+	if themePath == "" {
+		return fmt.Errorf("themePath is empty")
+	}
+	psScript := fmt.Sprintf(`
+# Set theme path
+Set-ItemProperty -Path "HKCU:\\Software\\Microsoft\\Windows\\CurrentVersion\\Themes" -Name "CurrentTheme" -Value "%s"
+
+# Set slideshow options
+Set-ItemProperty -Path "HKCU:\\Control Panel\\Desktop" -Name "SlideshowEnabled" -Value 1
+Set-ItemProperty -Path "HKCU:\\Control Panel\\Personalization\\Desktop Slideshow" -Name "SlideshowEnabled" -Value 1
+Set-ItemProperty -Path "HKCU:\\Control Panel\\Desktop\\PerMonitorSettings" -Name "SlideshowEnabled" -Value 1 -ErrorAction SilentlyContinue
+
+# Clear TranscodedWallpaper (sometimes holds stale data)
+$transcoded = "$env:APPDATA\\Microsoft\\Windows\\Themes\\TranscodedWallpaper"
+if (Test-Path $transcoded) { Remove-Item $transcoded -Force -ErrorAction SilentlyContinue }
+
+# Refresh system settings
+RUNDLL32.EXE user32.dll, UpdatePerUserSystemParameters
+
+# 🧠 Apply the .theme file again to force slideshow to start (silent, but effective)
+Start-Process -FilePath "%s" -WindowStyle Hidden
+
+Start-Sleep -Milliseconds 500
+`, themePath, themePath)
+
+	cmd := exec.Command("powershell", "-NoProfile", "-ExecutionPolicy", "Bypass", "-Command", psScript)
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		return fmt.Errorf("failed to apply theme: %w\nOutput: %s", err, string(output))
+	}
+
+	fmt.Fprintln(os.Stdout, "Theme applied output:", string(output))
+	return nil
+}
